radv: reduce radv_input_assembly_info::primitive_topology to 8-bit
It's large enough.

+++ src/amd/vulkan/radv_private.h
@@ -1882,7 +1882,7 @@ struct radv_vertex_input_info {
 };
 
 struct radv_input_assembly_info {
-   uint32_t primitive_topology;
+   uint8_t primitive_topology;
    bool primitive_restart_enable;
 };
